Extracted benchmarks into `benches` folder and introduced new ones for some intrinsics.
